Rename app.py to server.py
