+++ conftest.py
@@ -1,5 +1,7 @@
+"""Fixture configuration for pytest."""
 import pytest
-from app import create_app
+
+from server import create_app
 
 
 @pytest.fixture
